refactoring, sight cards folder
