+++ lib/ui/widgets/sight_cards/sight_card.dart
